Add utils.c

+++ utils.c
@@ -4,8 +4,8 @@ int _putchar(char c)
 {
     return (write(1, &c, 1));
 }
-    
-int print_number(int n)
+
+int print_number(long n)
 {
     int count = 0;
